requirements.txt
pycryptodome==3.9.7
tqdm==4.46.0
